modify to u64

+++ src/events.cairo
@@ -4,24 +4,24 @@ use starknet::ContractAddress;
     pub struct AccountCreated {
         pub name: ByteArray,
         pub address: ContractAddress,
-        pub balance: u128,
+        pub balance: u64,
     }
 
     #[derive(Clone, Drop, Debug, starknet::Event)]
     pub struct DepositMade {
-        pub amount: u128,
+        pub amount: u64,
         pub address: ContractAddress,
     }
 
     #[derive(Clone, Drop, Debug, starknet::Event)]
     pub struct WithdrawalMade {
-        pub amount: u128,
+        pub amount: u64,
         pub address: ContractAddress,
     }
 
     #[derive(Clone, Drop, Debug, starknet::Event)]
     pub struct TransferMade {
-        pub amount: u128,
+        pub amount: u64,
         pub from: ContractAddress,
         pub to: ContractAddress,
     }
@@ -30,5 +30,5 @@ use starknet::ContractAddress;
     pub struct AccountClosed {
         pub closed: ContractAddress,
         pub beneficiary: ContractAddress,
-        pub amount: u128,
+        pub amount: u64,
     }
